Part: Excise GeometryExtension from Geometry
